engineV2/requirements.txt
# ============================================
# Core Framework (Latest)
# ============================================
Flask==3.1.2
Flask-CORS==6.0.2
Werkzeug==3.1.4

# ============================================
# Data Validation (Latest)
# ============================================
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1

# ============================================
# Google AI (Latest)
# ============================================
google-generativeai==0.8.6

# ============================================
# OpenAI (Latest)
# ============================================
openai==1.59.3

# ============================================
# LangGraph & LangChain (Latest)
# ============================================
langgraph==1.0.5
langchain-core==1.2.5

# ============================================
# PDF Processing (Latest)
# ============================================
pypdf==6.5.0

# ============================================
# HTTP & Requests (Latest)
# ============================================
requests==2.32.5
urllib3==2.6.2

# ============================================
# Utilities (Latest)
# ============================================
python-dateutil==2.9.0.post0
pytz==2025.2
click==8.3.1

# ============================================
# Testing (Latest)
# ============================================
pytest==9.0.2
pytest-cov==7.0.0
pytest-mock==3.15.1
pytest-flask==1.3.0

# ============================================
# Code Quality (Latest)
# ============================================
black==25.12.0
flake8==7.3.0
mypy==1.19.1

# ============================================
# Development (Latest)
# ============================================
ipython==9.8.0

# ============================================
# Windows Production Server (Latest)
# ============================================
waitress==3.0.2
